Move mock to separate file
