+++ neurobooth_os/mock/__init__.py
@@ -1,2 +1,3 @@
 from .mock_database import insert_mock_rows, delete_mock_rows
 from .mock_servers import mock_server_ctr, mock_server_acq, mock_server_stm
+from .gui import MockGUIElement, MockWindow
